refactor: move credential/view.py to credential/view_v1.py to ease future removal
Relates to JIRA: DISCOVERY-797
