+++ quipucords/api/views.py
@@ -3,7 +3,7 @@
 # ruff: noqa: F401
 
 from api.aggregate_report.view import aggregate_report
-from api.credential.view import CredentialViewSetV1, credential_bulk_delete
+from api.credential.view_v1 import CredentialViewSetV1, credential_bulk_delete
 from api.deployments_report.view import deployments
 from api.details_report.view import details
 from api.insights_report.view import insights
